RT-7 Fixed code smells

+++ src/main/java/com/github/szilex94/edu/round_tracker/rest/jakarta/annotation/not_zero/NotZeroIntValidator.java
@@ -7,6 +7,6 @@ public class NotZeroIntValidator implements ConstraintValidator<NotZero, Integer
 
     @Override
     public boolean isValid(Integer value, ConstraintValidatorContext context) {
-        return value.intValue() != 0;
+        return value != 0;
     }
 }
